Add configs for Gen2->3 conversion.
These configs supersede the previous "convertRepo.py". It is impossible
to convert an ingested Gen 2 dataset to Gen 3 in one pass, because of
how templates are handled in Gen 2, and trying to reconcile the dataset
conversion script's very specific portability requirements with a single
"almost works" config proved too difficult.